move to-array-processor.h out of runtime-common
