+++ runtime-common/stdlib/visitors/common-visitors-methods.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "runtime-common/stdlib/array/to-array-processor.h"
+class ToArrayVisitor;
 
 struct CommonDefaultVisitorMethods {
   // for f$instance_to_array(), f$to_array_debug()
