+++ runtime/exception.h
@@ -10,10 +10,10 @@
 #include "common/wrappers/string_view.h"
 #include "runtime-common/core/class-instance/refcountable-php-classes.h"
 #include "runtime-common/core/runtime-core.h"
-#include "runtime-common/stdlib/array/to-array-processor.h"
 #include "runtime/dummy-visitor-methods.h"
 #include "runtime/instance-copy-processor.h"
 #include "runtime/memory_usage.h"
+#include "runtime/to-array-processor.h"
 
 array<array<string>> f$debug_backtrace();
