+++ runtime/rpc.h
@@ -9,7 +9,7 @@
 #include "common/algorithms/hashes.h"
 #include "common/kprintf.h"
 #include "runtime-common/core/runtime-core.h"
-#include "runtime-common/stdlib/array/to-array-processor.h"
+#include "runtime/to-array-processor.h"
 #include "runtime/dummy-visitor-methods.h"
 #include "runtime/net_events.h"
 #include "runtime/resumable.h"
